Disable etag generation in radicale-remind for this adapter
Generated Etags include the dtstamp and thus would change whenever the
Remind input file was changed. As commented, we don't rely on etags
anyhow.

An alternative would be to change the UID to something static (the file
line, for example) and move the currently used hash into the etag. The
downside would be that whenever one line was added or deleted, all lines
below would get a new UID and be resynced. Even more, Unrelated changes
in different lines on both sides would result in not syncable changes as
the UID would not be local anymore.