Replace the href_is_etag by a proper etag generator
There are reminders which could change without the UID (generated by
Remind) changes. An example would be:

REM MSG today is [today()]

Now in such cases only the etag changes. But the general comment that
whenever the Remind line itself was changed, it get's a new UID still
holds true.