Support INFO tag in Remind 05.03.00 (behaviour change)
Drop the old %"summary at location%" description.

Old files can be converted with:

sed -E -e \
  's/(.*) MSG %"(.*) at (.*)%"(%_)?(.*)/\1 INFO "Location: \3" INFO "Description: \5" MSG \2/' \
  -e 's/(.*) MSG %"(.*)%"(%_)?(.*)/\1 INFO "Description: \4" MSG \2/' \
  -e 's/(.*) MSG (.*) at (.*)/\1 INFO "Location: \3" MSG \2/' \
  -e 's/%_/\\\n/g'

Note that this is an incomplete conversion and the user may need to fix
up resulting problems.

This also means that text outside of a %"...%" sequences is ignored.

ics2rem drops the --locations and --sep as they no longer make sense.