Use INFO string for location if it is an url
When converting from ics to rem and the location of an event is an url,
write it as an INFO string. E.g. (split over multiple lines for
clarity):

REM Sep 25 2025 AT 10:00 DURATION 1:00 TAG PUBLIC \
	INFO "url: https://example.com/" \
	MSG %"Message%"%_Description

This is useful for when using tkremind, as in this way it allows to open
the link with xdg-open by middle-clicking the event [1].

For completeness, the opposite is also done when converting from rem to
ics. However a location specified with the " at <location>" syntax in
the MSG will take precedence over the INFO string if both are present.

[1] https://man.archlinux.org/man/tkremind.1.en#HYPERLINKS